move vault
