+++ token/sdk/vault/vault.go
@@ -3,7 +3,7 @@ Copyright IBM Corp. All Rights Reserved.
 
 SPDX-License-Identifier: Apache-2.0
 */
-package network
+package vault
 
 import (
 	"sync"
